fix(gboost): Fix run-fn command to allow TS files
