+++ .changeset/tough-meals-shop.md
@@ -1,0 +1,5 @@
+---
+"gboost": patch
+---
+
+Fix run-fn command to allow TS files
